+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.1.6
 
+- Fixed memory leak in KFCConfiguration
+- Improved NVS support for KFCConfiguration
 - Generic API to access IRAM and DRAM (requires `PIO_FRAMEWORK_ARDUINO_MMU_CACHE16_IRAM48_SECHEAP_SHARED` for the ESP8266)
 - Rendering forms uses IRAM if available